safari-600.7-branch/LayoutTests/fast/images/animated-gif-body-delegated-background-image-expected.txt

Make sure the background image of the body element is animated if the body has 0 height and the background is delegated to the root

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bodyHeight is "0px"
PASS isBackgroundAnimated is true
PASS successfullyParsed is true

TEST COMPLETE
